Implemented vote validation checks
Submitted votes are now checked to ensure they have a token that can be turned into an email. If the vote is valid it is recorded in the 'Validated Results' sheet, otherwise the invalid vote is recorded in the 'Invalid Results' sheet.

In either case the user is sent an email informing them of how the vote was handled.

If the vote was invalid a further email is sent to all Editors and Viewers of the Results sheet telling them that manual results collection is required, and all the sheets are colored red